requirements.txt
google-genai>=1.12.0
pymupdf>=1.24.10
requests>=2.32.0
python-dotenv>=1.0.1
edge-tts>=7.0.0
